Containers/Spark/requirements.txt
cassandra-driver==3.28.0
requests==2.31.0
requests-toolbelt==1.0.0